Add multiple builders
