+++ src/builders/build.rs
@@ -1,0 +1,5 @@
+use crate::contract::{Contract};
+
+pub trait Build {
+    fn build(&self, directory: &str) -> Result<(String, Vec<Contract>), Box<dyn std::error::Error>>;
+}
