+++ src/builders/hardhat.rs
@@ -1,0 +1,11 @@
+use crate::builders::build::Build;
+use crate::contract::Contract;
+
+pub struct HardhatBuilder;
+/*
+impl Build for HardhatBuilder {
+    fn build(&self, directory: &str) -> Result<(String, Vec<Contract>), Box<dyn std::error::Error>> {
+
+    }
+}
+*/
