+++ src/cli.rs
@@ -1,8 +1,8 @@
-use crate::html_parsing::{WebsiteParser};
+use crate::parsers::parse::{WebsiteParser};
 use crate::parsers::code4rena::{Code4renaParser};
 use crate::parsers::sherlock::{SherlockParser};
 use crate::github_api;
-use crate::contract::{process_repository, process_out_directory, ContractKind};
+use crate::contract::{process_repository, ContractKind};
 use clap::Parser;
 use log;
 
@@ -39,32 +39,37 @@ impl Cli {
                         match github_api::clone_repository(&repo.url, &repo.name) {
                             Ok(_) => {
                                 log::info!("Repo cloned from {} to {}", &repo.url, &repo.name);
-                                process_repository(&repo.name);
-                                let (repo_name, contract_data) = process_out_directory(&repo.name);
-
-                                let mut sorted_contracts = contract_data;
-                                sorted_contracts.sort_by_key(|contract| match contract.contract_kind {
-                                    ContractKind::Interface => 0,
-                                    ContractKind::Contract => 1,
-                                });
-
-                                // Enumerate the Vec<Contract> received by calling process_out_directory
-                                for contract in sorted_contracts {
-                                    println!("Repository: {}", repo_name);
-                                    println!("Contract Name: {}", contract.contract_name);
-                                    match &contract.imports {
-                                        Some(imports) => println!("Number of imports: {}", imports.len()),
-                                        None => println!("Number of imports: 0"),
+                                
+                                match process_repository(&repo.name) {
+                                    Ok((repo_name, contract_data)) => {
+                                        let mut sorted_contracts = contract_data;
+                                        sorted_contracts.sort_by_key(|contract| match contract.contract_kind {
+                                            ContractKind::Interface => 0,
+                                            ContractKind::Contract => 1,
+                                        });
+        
+                                        // Enumerate the Vec<Contract> received by calling process_out_directory
+                                        for contract in sorted_contracts {
+                                            println!("Repository: {}", repo_name);
+                                            println!("Contract Name: {}", contract.contract_name);
+                                            match &contract.imports {
+                                                Some(imports) => println!("Number of imports: {}", imports.len()),
+                                                None => println!("Number of imports: 0"),
+                                            }
+                                            match contract.contract_kind {
+                                                ContractKind::Interface => {
+                                                    println!("Contract Type: Interface");
+                                                    Self::print_bytecode(contract.bytecode, args.truncate);
+                                                }
+                                                ContractKind::Contract => {
+                                                    println!("Contract Type: Contract");
+                                                    Self::print_bytecode(contract.bytecode, args.truncate);
+                                                }
+                                            }
+                                        }
                                     }
-                                    match contract.contract_kind {
-                                        ContractKind::Interface => {
-                                            println!("Contract Type: Interface");
-                                            Self::print_bytecode(contract.bytecode, args.truncate);
-                                        }
-                                        ContractKind::Contract => {
-                                            println!("Contract Type: Contract");
-                                            Self::print_bytecode(contract.bytecode, args.truncate);
-                                        }
+                                    Err(err) => {
+                                        eprintln!("Error processing repository: {}", err);
                                     }
                                 }
                             }
